♻️ Make requireUser the session-only page gate
lib/auth.ts already had a session-trusting requireUser; the previous
commit unknowingly shadowed it with a database-hitting duplicate. The
original moves to features/user/loaders.ts as the canonical gate: reads
trust the session cookie cache, mutations verify against the database.

getCurrentUser remains the database-verified read for callers that need
it: the safe-action client, api-key actions, the admin SSR helper and
admin-setup (stale role), and the profile page (deletedAt is absent
from the session snapshot).
