+++ apps/web/src/app/[locale]/(space)/settings/profile/page.tsx
@@ -1,4 +1,5 @@
 import type { Metadata } from "next";
+import { redirect } from "next/navigation";
 import { Suspense } from "react";
 
 import type { Params } from "@/app/[locale]/types";
@@ -17,9 +18,11 @@ import {
   SettingsPageHeader,
   SettingsPageTitle,
 } from "@/components/settings-layout";
-import { requireUser } from "@/features/user/loaders";
+import { getCurrentUser } from "@/features/user/loaders";
 import { Trans } from "@/i18n/client";
 import { getTranslation } from "@/i18n/server";
+import { getPathname } from "@/lib/pathname";
+import { buildSafeRedirectUrl } from "@/lib/utils/redirect";
 import {
   AccountDeletionSummary,
   AccountDeletionSummarySkeleton,
@@ -30,7 +33,18 @@ import { ProfileEmailAddress } from "./components/profile-email-address";
 import { ProfileSettings } from "./components/profile-settings";
 
 export default async function Page() {
-  const user = await requireUser();
+  // Read from the database — the pending deletion notice depends on
+  // deletedAt, which the session snapshot doesn't carry.
+  const user = await getCurrentUser();
+
+  if (!user) {
+    redirect(
+      buildSafeRedirectUrl({
+        destination: "/login",
+        returnUrl: await getPathname(),
+      }),
+    );
+  }
 
   return (
     <SettingsPage>
