+++ apps/web/src/app/[locale]/admin-setup/page.tsx
@@ -11,15 +11,26 @@ import {
   EmptyStateTitle,
 } from "@/components/empty-state";
 import { isInitialAdmin } from "@/features/instance-settings/utils";
-import { requireUser } from "@/features/user/loaders";
+import { getCurrentUser } from "@/features/user/loaders";
 import { Trans } from "@/i18n/client";
 import { getTranslation } from "@/i18n/server";
+import { getPathname } from "@/lib/pathname";
+import { buildSafeRedirectUrl } from "@/lib/utils/redirect";
 import { MakeMeAdminButton } from "./make-me-admin-button";
 
 export default async function AdminSetupPage() {
   // Read the role from the database — the session cookie cache can hold
   // a stale role.
-  const user = await requireUser();
+  const user = await getCurrentUser();
+
+  if (!user) {
+    redirect(
+      buildSafeRedirectUrl({
+        destination: "/login",
+        returnUrl: await getPathname(),
+      }),
+    );
+  }
 
   if (user.role === "admin") {
     redirect("/control-panel");
