+++ apps/web/src/features/user/loaders.ts
@@ -3,7 +3,8 @@ import "server-only";
 import { redirect } from "next/navigation";
 import { cache } from "react";
 import { getUser } from "@/features/user/data";
-import { getSession } from "@/lib/auth";
+import type { UserDTO } from "@/features/user/schema";
+import { getSession, getSessionState } from "@/lib/auth";
 import { InvalidSessionError } from "@/lib/errors/invalid-session-error";
 import { getPathname } from "@/lib/pathname";
 import { buildSafeRedirectUrl } from "@/lib/utils/redirect";
@@ -31,21 +32,33 @@ export const getCurrentUser = cache(async () => {
 });
 
 /**
- * The current signed-in user for pages that require auth: redirects to
- * /login when there is no session or the user is a guest, so callers
- * always get a user back. Use getCurrentUser where the user is optional.
+ * Gate for server pages that require a logged-in (non-guest) user.
+ * Trusts the session cookie cache (no database read). Redirects to /login
+ * when unauthenticated, fails the render on an unreadable session so a
+ * transient session-store failure can't feed a / ↔ /login redirect loop.
+ * Use getCurrentUser where the user is optional or must be read from the
+ * database (mutation gates, stale-role-sensitive pages).
  */
-export const requireUser = async () => {
-  const user = await getCurrentUser();
+export const requireUser = cache(async (): Promise<UserDTO> => {
+  const state = await getSessionState();
 
-  if (!user) {
+  if (state.status === "error") {
+    throw new Error("Failed to read session");
+  }
+
+  const user =
+    state.status === "authenticated" ? state.session.user : undefined;
+
+  if (!user || user.isGuest) {
+    const pathname = await getPathname();
     redirect(
-      buildSafeRedirectUrl({
-        destination: "/login",
-        returnUrl: await getPathname(),
-      }),
+      buildSafeRedirectUrl({ destination: "/login", returnUrl: pathname }),
     );
   }
 
+  if (user.banned) {
+    throw new InvalidSessionError();
+  }
+
   return user;
-};
+});
