+++ apps/web/src/trpc/server/create-ssr-helper.ts
@@ -3,8 +3,8 @@ import { notFound, redirect } from "next/navigation";
 import { cache } from "react";
 import superjson from "superjson";
 import { isInitialAdmin } from "@/features/instance-settings/utils";
-import { getCurrentUser } from "@/features/user/loaders";
-import { getSession, requireUser } from "@/lib/auth";
+import { getCurrentUser, requireUser } from "@/features/user/loaders";
+import { getSession } from "@/lib/auth";
 import { getPathname } from "@/lib/pathname";
 import { buildSafeRedirectUrl } from "@/lib/utils/redirect";
 import type { TRPCContext } from "../context";
